requirements.txt
# Core
openai>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# API
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0

# Test
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
